# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FitPaxPro Gym API",
        "description": "Gym module mobile authentication endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/fitpaxpro/FitPaxPro/public",
            "description": "Application Server"
        }
    ],
    "paths": {
...
        "/api/admin/gym-operations/gyms": {
            "get": {
                "tags": [
                    "Gym Operations"
                ],
                "summary": "List gyms",
                "description": "Returns paginated gym records for admin operations.",
                "operationId": "8fe5a95caa19c2ecc75abbd5b417040d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gym list fetched"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden: non-admin token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/gym-operations/attendance": {
            "get": {
                "tags": [
                    "Gym Operations"
                ],
                "summary": "List attendance logs",
                "description": "Returns paginated attendance logs for admin operations.",
                "operationId": "e2a3b9437a431bcc74deb484ed6d4dd1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendance list fetched"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden: non-admin token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/gym-operations/enquiries": {
            "get": {
                "tags": [
                    "Gym Operations"
                ],
                "summary": "List gym enquiries",
                "description": "Returns paginated gym enquiries for admin operations.",
                "operationId": "07cba8a1911a487dd3700b7d82eaf23d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enquiry list fetched"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden: non-admin token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/gym-operations/fee-plans": {
            "get": {
                "tags": [
                    "Gym Operations"
                ],
                "summary": "List gym fee plans",
                "description": "Returns paginated gym fee plans for admin operations.",
                "operationId": "6063860caa9d8b97bb0bfdd70a967061",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fee plan list fetched"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden: non-admin token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/gym-operations/gallery-media": {
            "get": {
                "tags": [
                    "Gym Operations"
                ],
                "summary": "List gym gallery media",
                "description": "Returns paginated gym gallery media records for admin operations.",
                "operationId": "8dffe9e3928ede60f2bb441d6b54428e",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gallery media list fetched"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden: non-admin token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/gym-operations/reviews": {
            "get": {
                "tags": [
                    "Gym Operations"
                ],
                "summary": "List gym reviews",
                "description": "Returns paginated gym reviews for admin operations.",
                "operationId": "e215a1d20ea9685397ba62c710a4d126",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review list fetched"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden: non-admin token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}